feat(interface): separated header from app.jsx
