+++ to-do/src/components/header/Header.css
@@ -1,0 +1,5 @@
+header{
+    display: flex;
+    width: 100%;
+    justify-content: space-evenly;
+}
