+++ to-do/src/components/header/Header.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import Search from '../search/Search'
+import Filter from '../filter/Filter'
+import './Header.css'
+export default function Header() {
+  return (
+    <header>
+        <div className='logo'></div>
+        <Search/>
+        <Filter/>
+    </header>
+  )
+}
