+++ to-do/src/components/taskForm/TaskForm.js
@@ -2,6 +2,7 @@ import React, {useState}  from 'react'
 import Button from '../button/Button'
 import './TaskForm.css'
 
+
 const TaskForm = ({addTask}) => {
     // useState para controlar a desrição e a categpria da tarefa
     // Valor inicial: vazio ('')
