+++ to-do/src/components/wrapper/Wrapper.css
@@ -1,0 +1,5 @@
+.wrapper_content{
+    height: 100vh;
+    width: 95vw;
+    background: #F9F9F9;
+}
