+++ to-do/src/index.css
@@ -5,11 +5,6 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  padding: 30px;
-  color: #333;
-  background-image: url('./img/background-body.jpg');
-  background-position: center;
-  background-size: cover;
 
 }
